created chomp branch

+++ controller.runmean.cc
@@ -137,7 +137,7 @@ double Controller::estimateParameters() {
 
 
 
-int Controller::chompWindow(int cint, double cwindDL) {
+int Controller::chompWindow(unsigned int cint, double cwindDL) {
   if (networkDown) return 0;
   uint64_t tStamp = timestamp();
   // if we have a zero congestion window, push it out of this regime
@@ -161,7 +161,7 @@ int Controller::chompWindow(int cint, double cwindDL) {
     //fprintf(fsend, "%lu: unseen last timestamp %lu = %lu\n", tStamp, lastAck, tStamp - lastAck );
     cint = 1;//cint/2;
   }
-  else if ((lastAck > 0)  && (cint == (int)lastCW)) {
+  else if ((lastAck > 0)  && (cint == lastCW)) {
     cint += 1; // don't ever stay in a state without exploring up
   }
   // // make sure %change in cint isn't too spiky : causes delays
